requirements.txt
opencv-python==4.6.0.66
Pillow==9.2.0
pyserial==3.5
streamlit==1.11.0
tensorflow==2.10.0
torchvision==0.13.1
